Share capture parsing and counter validation
